deduplicate extension.ts & unsupervised.ts files for both desktop and web
the extension.ts files were getting messy
